requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django==4.0.1
django-cors-headers==3.10.1
django-summernote==0.8.20.0
djangorestframework==3.13.1
packaging==21.3
psycopg2
psycopg2-binary
gunicorn
Pillow==9.0.0
pyparsing==3.0.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1